+++ NEWS.md
@@ -2,7 +2,7 @@
 
 #### New Features
 
-* Adds `cloud_config()` to launch remote daemons using a cloud / cloud-based managed platform. Currently supports Posit Workbench.
+* Adds `posit_workbench_config()` to launch remote daemons using the default-configured Posit Workbench Launcher.
 * New synchronous mode: `daemons(sync = TRUE)` causes mirai to run synchronously within the current process.
   This facilitates testing and debugging, e.g. via interactive `browser()` instances (#439).